+++ filtering/filter.go
@@ -185,20 +185,18 @@ func transformInPlaceV2(block *pbcodec.Block, include, exclude, systemActions *C
 	for _, trxTrace := range trxTraces {
 		trxTraceAddedToFiltered := false
 		trxTraceExcluded := true
-		var trxTop5Actors []string // per transaction
-		getTrxTop5Actors := func() []string {
-			if trxTop5Actors == nil {
-				trxTop5Actors = getTop5ActorsForTrx(trxTrace)
-			}
-			return trxTop5Actors
-		}
+
+		memoizableTrxTrace := &MemoizableTrxTrace{TrxTrace: trxTrace}
 
 		for _, actTrace := range trxTrace.ActionTraces {
-			if wasFiltered && !actTrace.FilteringMatched {
-				continue
+			if wasFiltered {
+				if !actTrace.FilteringMatched { // skip previously excluded
+					continue
+				}
+				actTrace.FilteringMatched = false // expected default to be false
 			}
 
-			passes, isSystem := shouldProcess(trxTrace, actTrace, getTrxTop5Actors, include, exclude, systemActions)
+			passes, isSystem := shouldProcess(memoizableTrxTrace, actTrace, include, exclude, systemActions)
 			if !passes {
 				continue
 			}
@@ -218,19 +216,16 @@ func transformInPlaceV2(block *pbcodec.Block, include, exclude, systemActions *C
 		}
 
 		if trxTrace.FailedDtrxTrace != nil {
-			trxTop5Actors = nil
-			getTrxTop5Actors = func() []string {
-				if trxTop5Actors == nil {
-					trxTop5Actors = getTop5ActorsForTrx(trxTrace.FailedDtrxTrace)
-				}
-				return trxTop5Actors
-			}
+			memoizableFailedDTrxTrace := &MemoizableTrxTrace{TrxTrace: trxTrace.FailedDtrxTrace}
 			for _, actTrace := range trxTrace.FailedDtrxTrace.ActionTraces {
-				if wasFiltered && !actTrace.FilteringMatched {
-					continue
+				if wasFiltered {
+					if !actTrace.FilteringMatched { // skip previously excluded
+						continue
+					}
+					actTrace.FilteringMatched = false // expected default to be false
 				}
 
-				passes, isSystem := shouldProcess(trxTrace.FailedDtrxTrace, actTrace, getTrxTop5Actors, include, exclude, systemActions)
+				passes, isSystem := shouldProcess(memoizableFailedDTrxTrace, actTrace, include, exclude, systemActions)
 				if !passes {
 					continue
 				}
@@ -322,19 +317,19 @@ func combineFilters(prev string, next *CELFilter) string {
 	return fmt.Sprintf("%s;;;%s", prev, next.code)
 }
 
-func shouldProcess(trxTrace *pbcodec.TransactionTrace, actTrace *pbcodec.ActionTrace, trxTop5ActorsGetter func() []string, include, exclude, systemActions *CELFilter) (pass bool, isSystem bool) {
-	activation := actionTraceActivation{trace: actTrace, trxScheduled: trxTrace.Scheduled, trxActionCount: len(trxTrace.ActionTraces), trxTop5ActorsGetter: trxTop5ActorsGetter}
+func shouldProcess(trxTrace *MemoizableTrxTrace, actTrace *pbcodec.ActionTrace, include, exclude, systemActions *CELFilter) (pass bool, isSystem bool) {
+	activation := NewActionTraceActivation(actTrace, trxTrace, "")
 	// If the include program does not match, there is nothing more to do here
-	if !include.match(&activation) {
-		if systemActions.match(&activation) {
+	if !include.match(activation) {
+		if systemActions.match(activation) {
 			return true, true
 		}
 		return false, false
 	}
 
 	// At this point, the inclusion expr matched, let's check it was included but should be now excluded based on the exclusion filter
-	if exclude.match(&activation) {
-		if systemActions.match(&activation) {
+	if exclude.match(activation) {
+		if systemActions.match(activation) {
 			return true, true
 		}
 		return false, false